updated flex

+++ Modulo3/Flex/CarStand/CarStand/Models/Inspecoes.cs
@@ -8,11 +8,11 @@ namespace CarStand.Models
 {
     public  class Inspecoes
     {
-        public int ID { get; private set; }
+        public long ID { get; private set; }
         public DateTime dataInspecao { get; private set; }
 
         public bool Resultado { get; private set; }
-        public Inspecoes(int id, DateTime datainspecao, bool resultado)
+        public Inspecoes(long id, DateTime datainspecao, bool resultado)
         {
             ID = id;
             dataInspecao = datainspecao;
